refactor: move norm balls to lp_norm_balls.py on shared projections
git mv norm_balls.py -> lp_norm_balls.py. Move the proximal operator into
the _NormBall base (projection onto the ball), with subclasses binding a
_norm and a _projection delegating to lp_norm_helpers. LInfNormBall is now
a _NormBall subclass (was a Box subclass); its projection is the elementwise
clamp, so behavior is unchanged. Update the atoms package imports.
